ProfilePage update andd custom hook
